+++ modules/core/src/main/scala/io/bartholomews/fsclient/core/FsApiClient.scala
@@ -1,9 +1,0 @@
-package io.bartholomews.fsclient.core
-
-import io.bartholomews.fsclient.core.http.FsClientSttpExtensions
-import io.bartholomews.fsclient.core.oauth.Signer
-import sttp.client3.SttpBackend
-
-abstract case class FsApiClient[F[_], S <: Signer](client: FsClient[F, S]) extends FsClientSttpExtensions {
-  val backend: SttpBackend[F, Any] = client.backend
-}