+++ modules/circe/src/main/scala/io/bartholomews/fsclient/circe/FsClientCirceApi.scala
@@ -57,7 +57,7 @@ trait FsClientCirceApi extends SttpCirceApi {
       accessToken,
       tokenType,
       expiresIn,
-      refreshToken,
+      Some(refreshToken),
       scope
     )