+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 - `FsApiClient`
 ### Changed
-- `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`, and has `refreshToken` required
+- `AccessTokenSigner` using `currentTimeMillis` instead of `nanoTime`
 - `baseRequest` taking an `userAgent` instead of `client`
 ### Fixed